requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
Flask-APScheduler
python-dotenv
requests